requirements.txt
fastapi
uvicorn
torch
tensorflow
numpy
pillow
python-multipart  # Untuk upload file
requests
gunicorn